check for create form 2

+++ src/Components/LoanBeneficiary/LoanBenListTable.jsx
@@ -6,6 +6,7 @@ import LoanDetailModal from "./LoanDetailModal";
 import { formatDate } from "../../hooks/formatDate";
 import EditLoanSidebar from "./EditLoanSidebar";
 import { deleteLoanBeneficiary } from "../../redux/slices/deleteLoanBeneficiarySlice";
+import Swal from "sweetalert2";
 
 const LoanBenListTable = () => {
   const dispatch = useDispatch();
@@ -44,8 +45,26 @@ const LoanBenListTable = () => {
   }
 
   const handelDelete = (id) => {
-    dispatch(deleteLoanBeneficiary(id));
-    dispatch(fetchLoanBeneList());
+
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        dispatch(deleteLoanBeneficiary(id));
+        dispatch(fetchLoanBeneList(state.loanBeneList.currentPage));
+        Swal.fire(
+          'Deleted!',
+          'Your file has been deleted.',
+          'success'
+        )
+      }
+    })
   };
 
   return (
